Revert new files.
